--- Sources/Spine/Model/Animation/Keyframes/Bone/BoneKeyframeModel.swift
@@ -1,10 +1,0 @@
-//
-//  BoneKeyframeModel.swift
-//  
-//
-//  Created by Max Gribov on 12.11.2022.
-//
-
-import Foundation
-
-protocol BoneKeyframeModel: CurvedKeyframeModel {}
