+++ Sources/Spine/Model/Animation/Keyframes/Bone/BoneKeyframeShearModel.swift
@@ -7,7 +7,7 @@
 
 import Foundation
 
-struct BoneKeyframeShearModel: BoneKeyframeModel {
+struct BoneKeyframeShearModel: CurvedKeyframeModel {
     
     let time: TimeInterval
     let shear: CGVector
