+++ Sources/Spine/Model/Animation/Keyframes/EventKeyfarameModel.swift
@@ -7,7 +7,7 @@
 
 import SpriteKit
 
-struct EventKeyfarameModel: KeyframeModel, AnimationGroupModel {
+struct EventKeyfarameModel: KeyframeModel {
     
     let time: TimeInterval
     let event: String
