--- Sources/Spine/Model/Animation/Keyframes/Slot/SlotKeyframeModel.swift
@@ -1,10 +1,0 @@
-//
-//  SlotKeyframeModel.swift
-//  
-//
-//  Created by Max Gribov on 12.11.2022.
-//
-
-import Foundation
-
-protocol SlotKeyframeModel: KeyframeModel {}
